+++ frontend/package.json
@@ -40,6 +40,14 @@
     ]
   },
   "devDependencies": {
-    "eslint": "^5.16.0"
+    "@types/flux": "^3.1.9",
+    "@types/jest": "^26.0.8",
+    "@types/node": "^14.0.27",
+    "@types/react": "^16.9.44",
+    "@types/react-dom": "^16.9.8",
+    "@types/react-helmet": "^6.0.0",
+    "@types/react-router-dom": "^5.1.5",
+    "eslint": "^5.16.0",
+    "typescript": "^3.9.7"
   }
 }
