+++ frontend/src/components/pages/Work.tsx
