+++ frontend/src/react-app-env.d.ts
@@ -1,0 +1,3 @@
+/// <reference types="react-scripts" />
+type BooleanSetter = (x: boolean) => void
+type NumberSetter = (x: number) => void
