+++ frontend/src/staticData.tsx
